move tests to lib file as suggested in https://github.com/sofiazcoaga/conways-game-of-life/pull/1#pullrequestreview-2888784949
